api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis[hiredis]>=5.0.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
httpx>=0.26.0
anthropic>=0.40.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
boto3>=1.34.0
aiofiles>=23.0.0
celery[redis]>=5.3.0
websockets>=12.0
pyyaml>=6.0.1
Pillow>=10.0.0
orjson>=3.9.0
slowapi>=0.1.9
